+++ bonsai-ai/tests/test_ping_pong.py
@@ -1,0 +1,15 @@
+def test_train_sim(train_sim, capsys):
+    """ 
+    Tests that pings are received by our mock server.
+    This test is contained in it's own file because it
+    captures stdout output and checks that a received
+    message was printed
+    """
+    train_sim.brain.config.ping_interval = 1.0
+    counter = 0
+    while train_sim.run():
+        if counter == 2000:
+            break
+        counter += 1
+    out, err = capsys.readouterr()
+    assert('Received PING: ' in out)